+++ src/interpretation/mod.rs
@@ -1,15 +1,15 @@
 mod command;
+mod interpretation_item;
 mod interpretation_stream;
 mod interpreted_stream;
 mod interpreter;
 mod interpreter_parse_stream;
-mod next_item;
 mod variable;
 
 pub(crate) use command::*;
+pub(crate) use interpretation_item::*;
 pub(crate) use interpretation_stream::*;
 pub(crate) use interpreted_stream::*;
 pub(crate) use interpreter::*;
 pub(crate) use interpreter_parse_stream::*;
-pub(crate) use next_item::*;
 pub(crate) use variable::*;